Move OperationSimplifier methods to cpp file
